+++ .gitignore
@@ -18,3 +18,10 @@ local.properties
 # Eclipse project files
 .classpath
 .project
+
+# IntelliJ
+.idea/
+*.iml
+
+# Maven
+target/
