+++ demo/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.sothree.slidinguppaneldemo"
+    package="com.sothree.slidinguppanel.demo"
     android:versionCode="1"
-    android:versionName="1.0" >
+    android:versionName="1.0.0">
 
     <uses-sdk
         android:minSdkVersion="11"
@@ -12,10 +12,10 @@
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
+        android:theme="@style/AppTheme">
         <activity
-            android:name="com.sothree.slidinguppaneldemo.DemoActivity"
-            android:label="@string/app_name" >
+            android:name="com.sothree.slidinguppanel.demo.DemoActivity"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
