+++ demo/res/drawable-hdpi/ic_launcher.png
