+++ demo/res/drawable-mdpi/ic_launcher.png
