+++ demo/res/drawable-xhdpi/graphic.png
