+++ demo/res/layout/activity_demo.xml
@@ -2,12 +2,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DemoActivity" >
+    tools:context=".DemoActivity">
 
-    <com.sothree.slidinguppaneldemo.SlidingUpPanelLayout
+    <com.sothree.slidinguppanel.SlidingUpPanelLayout
         android:id="@+id/sliding_layout"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" >
+        android:layout_height="match_parent">
 
         <TextView
             android:layout_width="match_parent"
@@ -20,7 +20,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="#eee"
-            android:orientation="vertical" >
+            android:orientation="vertical">
 
             <TextView
                 android:layout_width="match_parent"
@@ -35,18 +35,18 @@
                 android:layout_height="34dp"
                 android:gravity="center|top"
                 android:text="Brought to you by http://umanoapp.com"
-                android:textSize="14sp" 
-                android:autoLink="web"/>
-            
+                android:textSize="14sp"
+                android:autoLink="web" />
+
             <ImageView
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_weight="1"
                 android:src="@drawable/graphic"
-                android:scaleType="fitStart" >
-                
-            </ImageView>           
+                android:scaleType="fitStart">
+
+            </ImageView>
         </LinearLayout>
-    </com.sothree.slidinguppaneldemo.SlidingUpPanelLayout>
+    </com.sothree.slidinguppanel.SlidingUpPanelLayout>
 
 </RelativeLayout>
