+++ demo/res/menu/demo.xml
