+++ demo/res/values/dimens.xml
