+++ demo/res/values/strings.xml
