+++ library/res/drawable/above_shadow.xml
